Fix some warnings

+++ src/server/appserver.cpp
@@ -3943,7 +3943,7 @@ void AppServer::addError(const Error &error) {
                 }
             }
         }
-        for (int errorId: toBeRemovedErrorIds) {
+        for (int64_t errorId: toBeRemovedErrorIds) {
             bool found = false;
             if (!ParmsDb::instance()->deleteError(errorId, found)) {
                 LOG_WARN(Log::instance()->getLogger(), "Error in ParmsDb::deleteError");
